+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 ### Changed
+- Many generic methods that previously took a `Component` now instead take either a
+  `ComponentRef<'a>` or a `Query` to improve consistency with queries and address a common footgun:
+  - `World::get` and `EntityRef::get` now take shared or unique references to component types
 - `SerializeContext` traits now take their serializer arguments by value, and must call `end()`
   themselves.